Move PullRequest and ReviewRequest to GitHubApi
