--- Project Walking/review-police/src/model/PullRequest.ts
@@ -1,9 +1,0 @@
-import { ReviewRequest } from "./ReviewRequest";
-
-export interface PullRequest {
-    title: string,
-    url: string
-    number: number,
-    status: string,
-    reviewRequests: Array<ReviewRequest>
-}
