--- Project Walking/review-police/src/model/ReviewRequest.ts
@@ -1,6 +1,0 @@
-export interface ReviewRequest {
-    url: string,
-    reviewer: string
-    pullNumber: number,
-    title: string
-}
